do not remove all networks before adding a network in Connect function
This had to be done because of an issue encountered while testing the wifi Connect function on real hardware. The Connect function was removing all networks and then trying to add a network. Wpa_supplicant service on receiving the command for removing all networks was also bringing the interface down for some reason (not sure what the reason is) and then the command for adding a network was hence failing since the interface was not up